backend/requirements.txt
google-cloud-aiplatform==1.46.0